Refactoring : WARNNINGS
Pascal case, snake case 혼용한 컴포넌트 pascal case로 일괄 수정
