+++ src/component/AboutAdmin/SD_Account/CMS/List/CMSList.js
@@ -1,4 +1,6 @@
-export default function CMSList({sortedAccounts}) {
+import styles from './CMSList.module.css';
+
+export default function CMSList({ sortedAccounts }) {
 
 
     return (
